move contents of package zng/flattener to expr
